GPU Trellis SDM: fp64 DAS path produces clean stitching at 43 segments
- Fix stale ext_seed: DAS path passes NULL ext_seed (was using stale
  static pointer from previous experiments → kernel produced no output)
- gpu_cuda_trellis_das with fp64 input, 2-pass DAS, 43 segments:
  clean audio confirmed (no pops/crackling from stitching)
- Performance: 1.77x RT at DSD64 (over real-time → stuttering)
- Root cause of slowness: 2-pass × 43 segments × fp64 × 2 channels
- All GPU trellis functions now accept fp64 input (const double *)
- ensure_sdm_bufs allocates d_sdm_in as double (covers both fp64 trellis
  and fp32 precorr)
- Next: optimize performance to sub-1x RT

+++ include/gpu_compute.h
@@ -211,7 +211,7 @@ int gpu_cuda_precorr_setup(void *ctx, int order,
                             const float *pred_table, float state_limit);
 
 /* Persistent-buffer SDM process (no per-chunk alloc) */
-int gpu_cuda_trellis(void *ctx, const float *in, float *out, size_t count);
+int gpu_cuda_trellis(void *ctx, const double *in, float *out, size_t count);
 /* DAS (Density-Aligned Stitching) GPU pipeline: SDM + stitch + assemble.
  * All channels processed simultaneously. Input/output: [num_ch × count] */
 int gpu_cuda_trellis_das(void *ctx, const double *in, float *out,
